Fix depth search does not change root and gives -1 not found always
The former was because I should not reference this.root every time
because it's left and right child does not change. Simply pass in a root
rather. We want the new root to change each time with each function call
and not stick to an original root. The latter was because of the
if statement for if null I originally put if node || root === null
return -1. Which is incorrect because that checks if node is present or
root is null then return -1 so it will always return -1.

Finally I short circuited the function end. I moved the checks for if
leftDepth or rightDepth > 0 then return without checking the next
branch. If the node is found in the left branch recursively you don't
need to return the right branch because it will always return -1.
If it found in the left branch the right branch will be -1 and vice
versa recursively. So you can return early without checking the next
branch.